refactor: data 레이어 패키지 구조 변경
- `com.mulkkam.util.work` 패키지를 `com.mulkkam.data.local.work`로 이동
- `com.mulkkam.data.remote.service.HealthService`를 `com.mulkkam.data.local.service.HealthService`로 이동
- 이에 따른 import 경로 수정
