+++ android/app/src/main/java/com/mulkkam/data/local/work/CalorieWorker.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.util.work
+package com.mulkkam.data.local.work
 
 import android.content.Context
 import android.util.Log
