+++ android/app/src/main/java/com/mulkkam/di/WorkInjection.kt
@@ -3,10 +3,10 @@ package com.mulkkam.di
 import android.content.Context
 import androidx.work.Configuration
 import androidx.work.WorkManager
+import com.mulkkam.data.local.work.CalorieSchedulerImpl
+import com.mulkkam.data.local.work.WorkerFactory
 import com.mulkkam.di.RepositoryInjection.healthRepository
 import com.mulkkam.domain.work.CalorieScheduler
-import com.mulkkam.util.work.CalorieSchedulerImpl
-import com.mulkkam.util.work.WorkerFactory
 
 object WorkInjection {
     private lateinit var workManager: WorkManager
